Print a count of the DBs to process, when run with --verbose
Before actually processing the databases, print how many there are
(in the same, existing line in which we print the actual list of DBs).
This gives us a convenient way to compare the number of DBs we had
hoped to process, on the one hand, with the number we actually
actually processed, on the other (the latter is already printed
at the end of the flexible_freeze.py run, absent exceptions).