Add an optional per-table lock timeout
Prior to this commit, if a long (auto)vacuum or other lock
was preventing Flexible Freeze from vacuuming a table,
it would wait for that lock indefinitely, or until
run_min minutes elapsed (if specified).

This commit addresses the problem by adding --lock-timeout,
an optional flag for specifying the timeout (in milliseconds).
If Flexible Freeze can't acquire a table lock after
that amount of time, it gives up and moves on to the next table.